be more careful on delete, only append tags

+++ split_video.sh
@@ -3,5 +3,10 @@
 SOURCE_FILE="$1"
 DEST_DIR=$2
 
+if [ "${DEST_DIR}" == "" ]; then
+  echo "You must give me a destination dir"
+  exit
+fi
+
 rm -rf ${DEST_DIR}/*.png
 ffmpeg -i "${SOURCE_FILE}"  -vf fps=1/2 -f image2 "${DEST_DIR}/frame-%07d.png"
